pyproject.toml
[tool.poetry]
name = "azfunc"
version = "0.1.0"
description = ""
authors = ["Niten Nashiki <niten.nashiki@lightz-inc.com>"]

[tool.poetry.dependencies]
python = "^3.9"
azure-functions = "^1.10.1"
pydantic = "^1.9.0"
az-evgrid-pydantic-schema = "^0.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.10.1"
black = "^22.3.0"
taskipy = "^1.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
  | \.git
  | \.mypy_cache
  | \.venv
  | _build
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"

[tool.taskipy.tasks]
black = "black azfunc && black tests"
isort = "isort azfunc && isort tests"
# poetry run task test
test = "pytest . -s"
# poetry run task fmt
fmt = { cmd = "task black && task isort", help = "format code" }

